+++ fiberplane-ui/src/components/Icon/svg.d.ts
@@ -1,0 +1,9 @@
+declare module "*.svg" {
+  import * as React from "react";
+
+  const component: React.FunctionComponent<
+    React.SVGProps<SVGSVGElement> & { title?: string }
+  >;
+
+  export default component;
+}