+++ fiberplane-ui/src/theme/ThemeProvider.tsx
@@ -1,0 +1,22 @@
+import { ThemeProvider as StyledComponentsThemeProvider } from "styled-components";
+
+import { extendedTheme } from "./extendedTheme/extendedTheme";
+import { GlobalStyle } from "./globalStyle";
+import "./styled-components.d";
+
+type ThemeProviderProps = {
+  children: React.ReactNode;
+};
+
+/**
+ * Fiberplane theme provider that includes the theme's CSS variables & typed
+ * theme object.
+ */
+export function ThemeProvider({ children }: ThemeProviderProps) {
+  return (
+    <StyledComponentsThemeProvider theme={extendedTheme}>
+      <GlobalStyle />
+      {children}
+    </StyledComponentsThemeProvider>
+  );
+}