+++ fiberplane-ui/src/theme/extendedTheme/index.ts
@@ -1,0 +1,2 @@
+export { spacingCssVariables } from "./spacing";
+export * from "./extendedTheme";